backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.0
python-multipart==0.0.6
openai==1.51.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aiofiles==23.2.1
asyncpg==0.29.0
email-validator==2.1.0
bcrypt==4.1.2
gunicorn==21.2.0 